requirements.txt
fastapi[standard]
fastapi-sqlalchemy==0.2.1
pydantic==2.5.3
uvicorn==0.27.0
python-dotenv==1.0.1
sqlalchemy==2.0.25
annotated-types==0.6.0
anyio==4.2.0
colorama==0.4.6
greenlet==3.0.3
h11==0.14.0
idna==3.6
jmespath==1.0.1
mako==1.3.1
markupsafe==2.1.4
pydantic-core==2.14.6
python-dateutil==2.8.2
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
typing-extensions==4.9.0
urllib3==2.0.7
python-multipart
psycopg2
redis
pyyaml
docker
rsa
Jinja2
bs4
celery
flower
fastapi-cli
apscheduler
unidecode
alembic
sqladmin[full]
itsdangerous